Turn init_replication() to a hook
The hook allow to override the default behavior of creating a standby
with pg_basebackup.
So we can implement creating a standby from a pgbackrest backup easily.

The hook is defined with firstresult=True, so the first hook returning
something else than None will stop the process of calling these hooks.

The default implementation, defined with trylast=True, is to initialize
replica with pg_basebackup.

Related to #134