Refactor start_program() as a context manager class
We turn start_program() into a Program class and add a context manager
protocol to it. This will allow callers to perform checks after program
startup, e.g. to verify if the service is running okay. This might be
useful for asynchronous services (e.g.  Patroni) which do not fail fast
but rather wait indefinitely for other services to be up. With this
context manager API, we'd typically check if the service is up and
eventually let the process be killed if we decide it's necessary.