Add a rmtree() method to Context
This method installs an error handler which emits a log message when
some file in the directory tree fails to get deleted. This is not a big
deal in general, and can occur for instance when the instance is stopped
asynchronously (e.g. through Patroni) and dropped at the same time.

The absolute_path() decorator in util module is dropped as it no longer
applies and we can in fact just repeat the logic twice.