Drop "values" parameter from instances.configure()
Instead we inspect the 'manifest' parameter for its configuration and
build the complete configuration with port and locale settings in
instances.configuration().

In functional tests, we drop configure_instance() helper function since
it's equally simple to call instances.configure() directly.

Changes in tests are mostly due to accounting for 'locale' settings in
the new behavior.