Compute overall configuration changes in instances.configure()
We move the individual 'changes' computation from configuration()
function to a global computation (based on postgresql.conf) in
configure() function.

This way, configuration() is only concerned about generating the
configuration (not about changes).

As a consequence, and as can be seen in unit tests update, the 'changes'
value now accounts for 'site.conf' file, which is less surprising for
the user.