Move out site_config() from a Context method to a plain function
The sole reason for having site_config() defined on the Context was to
mock it in tests. But the amount of code necessary to make things work
on application side was not quite reasonable.

Instead we now just define one behaviour, encoded in util.site_config(),
to retrieve configuration or template files from user/etc/dict
locations. And in tests, we "simply" mock this function.