Set default 'port' field value during validation
We remove the default value for Instance's port field and set it during
validation. In order this to work, we need a pre=True validator
otherwise we'd first get an error at field validation because the field
is required.

In addition to ensuring that configuration['port'] is consistent
with the 'port' field, we possibly convert the former into an integer.